Combine AddHotel and EditHotel forms
- Make a HotelForm component and use it for both forms
